chore: create src/story folder and move miscellaneous into story folder
